Improved file name generation and handling of duplicate titles
fix: generated file names for covers, screens and game files is improved to consider more characters to avoid that two games point to the same files.
fix: When adding a new game where the file names becomes equal to an existing game in the db an additional index is added to the file name as a suffix.
fix: file names are re-calculated when a game is saved and renamed if not matching the new rules.