Feat: Support for handling multiple disks
The manager can keep track of 5 extra disks (6 in total) for a game, listed on a new Tab in the UI. When scraping games or importing from a gamebase db available extra disks are added automatically.
Extra disks are added to a new folder named "extradisks". The extra disks are exported when a game is exported to file loader. Exporting to carousel will not include extra disks since the carousel have no disk swap support.